add debug message when interaction interval skips
When an xmpp notification message is skipped due to last_cli_interaction
not greater than the configured xir_notify_delay a message is now
printed to tell the user why the message was skipped. (issue #14)

Add version information to /xir-status page.

Fix some logic in sig_public_msg so that the interact_interval is always
explicitly set, as opposed to initializing it to zero, then only setting
it if the message was not to ournick. The semantics do not change, but
the section of code is more readable this way.